// crm_frontend_3/src/api/api2.ts
export interface Create_CPUDto {
    name_Сomponent: string ;
    price_CPU: number;
    quantity_CPU: number;
    number_Cores: number;
    purity_CPU: string ;
    name_Company: string;
}